Revert "[torch.compile] Fix RMSNorm + quant fusion in the non-cutlass-fp8 case, rename RedundantReshapesPass to NoopEliminationPass (#10902)"
This reverts commit bd56c983d6fe8ff93bddd5faaf8d96e01c90fd83.
